Profile: fix getdict ip lookup on 32-bit hosts
`convert(Ptr{Cvoid}, ip)` only has methods for word-size integers, so
`getdict` on `Vector{UInt64}` data threw a MethodError inside the
lookup task on 32-bit
(https://buildkite.com/julialang/julia-pr/builds/659#019f8702-fa80-4b50-b6c5-facde9ff897a).
Convert through `UInt` first.

+++ stdlib/Profile/src/Profile.jl
@@ -489,7 +489,8 @@ function getdict!(dict::LineInfoDict, data::Vector{<:Unsigned})
 end
 
 function _lookup_corrected(ip::Unsigned)
-    st = lookup(convert(Ptr{Cvoid}, ip))
+    # convert through UInt: only word-size integers convert directly to Ptr
+    st = lookup(convert(Ptr{Cvoid}, UInt(ip)))
     # To correct line numbers for moving code, put it in the form expected by
     # Base.update_stackframes_callback[]
     stn = map(x->(x, 1), st)
